add label to the cell itself so the function only has to be called on the initial parsing of the data
